[CIAS-3467] - Add terms accepted step
