+++ app/components/TermsCheckbox/styled.tsx
@@ -1,0 +1,10 @@
+// @ts-ignore
+import styled from 'styled-components';
+
+import Box from 'components/Box';
+
+export const TermsAndConditions = styled(Box)`
+  p {
+    margin: 10px 0;
+  }
+`;
