+++ app/containers/LoginPage/constants.js
@@ -1,2 +1,3 @@
 export const LOGIN_FORM_VIEW = 'LOGIN_FORM_VIEW';
 export const CODE_VERIFICATION_VIEW = 'CODE_VERIFICATION_VIEW';
+export const TERMS_NOT_ACCEPTED_VIEW = 'TERMS_NOT_ACCEPTED_VIEW';
