+++ app/global/reducers/auth/actions.js
@@ -43,6 +43,10 @@ import {
   FETCH_SELF_DETAILS_REQUEST,
   FETCH_SELF_DETAILS_SUCCESS,
   UPDATE_USERS_TIMEZONE,
+  TERMS_NOT_ACCEPTED,
+  TERMS_ACCEPT_REQUEST,
+  TERMS_ACCEPT_SUCCESS,
+  TERMS_ACCEPT_ERROR,
 } from './constants';
 
 export const logIn = (user) => actionBuilder(LOG_IN_USER, { user });
@@ -128,6 +132,9 @@ export const loginError = (error) => actionBuilder(LOGIN_ERROR, { error });
 export const verificationCodeNeeded = () =>
   actionBuilder(VERIFICATION_CODE_NEEDED, {});
 
+export const termsNotAccepted = (fields) =>
+  actionBuilder(TERMS_NOT_ACCEPTED, { fields });
+
 export const verificationCodeRequest = (verificationCode) =>
   actionBuilder(VERIFICATION_CODE_REQUEST, { verificationCode });
 
@@ -144,3 +151,9 @@ export const fetchSelfDetailsSuccess = (user) =>
 
 export const updateUsersTimezone = (timezone) =>
   actionBuilder(UPDATE_USERS_TIMEZONE, { timezone });
+
+export const termsAcceptRequest = (fields, onSuccess) =>
+  actionBuilder(TERMS_ACCEPT_REQUEST, { fields, onSuccess });
+export const termsAcceptSuccess = () => actionBuilder(TERMS_ACCEPT_SUCCESS, {});
+export const termsAcceptError = (error) =>
+  actionBuilder(TERMS_ACCEPT_ERROR, { error });
