+++ app/global/reducers/auth/constants.js
@@ -69,3 +69,9 @@ export const FETCH_SELF_DETAILS_REQUEST = 'app/auth/FETCH_SELF_DETAILS_REQUEST';
 export const FETCH_SELF_DETAILS_SUCCESS = 'app/auth/FETCH_SELF_DETAILS_SUCCESS';
 
 export const UPDATE_USERS_TIMEZONE = 'app/auth/UPDATE_USERS_TIMEZONE';
+
+export const TERMS_NOT_ACCEPTED = 'app/auth/TERMS_NOT_ACCEPTED';
+
+export const TERMS_ACCEPT_REQUEST = 'app/auth/TERMS_ACCEPT_REQUEST';
+export const TERMS_ACCEPT_SUCCESS = 'app/auth/TERMS_ACCEPT_SUCCESS';
+export const TERMS_ACCEPT_ERROR = 'app/auth/TERMS_ACCEPT_ERROR';
