+++ app/global/reducers/auth/selectors.js
@@ -34,6 +34,15 @@ export const makeSelectLoginFormData = () =>
 export const makeSelectVerificationNeeded = () =>
   createSelector(selectAuth, (authState) => authState.verificationCodeNeeded);
 
+export const makeSelectTermsNotAccepted = () =>
+  createSelector(selectAuth, (authState) => authState.termsNotAccepted);
+
+export const makeSelectTermsExtraFields = () =>
+  createSelector(
+    selectAuth,
+    (authState) => authState.termsNotAcceptedExtraFields,
+  );
+
 export const makeSelectVerificationSuccess = () =>
   createSelector(selectAuth, (authState) => authState.verificationCodeSuccess);
